+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Show certain data in the gameview only if they are available
 - Update Notification for new Versions in the non Microsoft Store Releases
 - Check for offline cache integrity
+- Uninstall games
 - Bug fix: Crash if you navigate to Settings/Data and never downloaded a game before
 - Bug fix: Image was requested although an invalid image id was specified
 - Bug fix: Game Title was also not displayed when the Release Date was not set